--- utils/gear-runtime-test-cli/src/mock.rs
@@ -1,188 +1,0 @@
-// This file is part of Gear.
-
-// Copyright (C) 2021 Gear Technologies Inc.
-// SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
-
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-// use crate as pallet_gear;
-use frame_support::traits::{FindAuthor, OnFinalize, OnIdle, OnInitialize};
-use frame_support::{construct_runtime, parameter_types};
-use frame_system as system;
-use sp_core::H256;
-use sp_runtime::{
-    testing::Header,
-    traits::{BlakeTwo256, IdentityLookup},
-};
-
-type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
-type Block = frame_system::mocking::MockBlock<Test>;
-type AccountId = u64;
-
-pub(crate) const USER_1: AccountId = 1000001;
-pub(crate) const USER_2: AccountId = 1000002;
-pub(crate) const USER_3: AccountId = 1000003;
-pub(crate) const LOW_BALANCE_USER: AccountId = 1000004;
-pub(crate) const BLOCK_AUTHOR: AccountId = 255;
-
-// Configure a mock runtime to test the pallet.
-construct_runtime!(
-    pub enum Test where
-        Block = Block,
-        NodeBlock = Block,
-        UncheckedExtrinsic = UncheckedExtrinsic,
-    {
-        System: system::{Pallet, Call, Config, Storage, Event<T>},
-        Gear: pallet_gear::{Pallet, Call, Storage, Event<T>},
-        GearDebug: pallet_gear_debug::{Pallet, Call, Storage, Event<T>},
-        Balances: pallet_balances::{Pallet, Call, Storage, Config<T>, Event<T>},
-        Authorship: pallet_authorship::{Pallet, Storage},
-        Timestamp: pallet_timestamp::{Pallet, Call, Storage, Inherent},
-    }
-);
-
-impl pallet_balances::Config for Test {
-    type MaxLocks = ();
-    type MaxReserves = ();
-    type ReserveIdentifier = [u8; 8];
-    type Balance = u128;
-    type DustRemoval = ();
-    type Event = Event;
-    type ExistentialDeposit = ExistentialDeposit;
-    type AccountStore = System;
-    type WeightInfo = ();
-}
-
-parameter_types! {
-    pub const BlockHashCount: u64 = 250;
-    pub const SS58Prefix: u8 = 42;
-    pub const ExistentialDeposit: u64 = 1;
-}
-
-impl system::Config for Test {
-    type BaseCallFilter = frame_support::traits::Everything;
-    type BlockWeights = ();
-    type BlockLength = ();
-    type DbWeight = ();
-    type Origin = Origin;
-    type Call = Call;
-    type Index = u64;
-    type BlockNumber = u64;
-    type Hash = H256;
-    type Hashing = BlakeTwo256;
-    type AccountId = AccountId;
-    type Lookup = IdentityLookup<Self::AccountId>;
-    type Header = Header;
-    type Event = Event;
-    type BlockHashCount = BlockHashCount;
-    type Version = ();
-    type PalletInfo = PalletInfo;
-    type AccountData = pallet_balances::AccountData<u128>;
-    type OnNewAccount = ();
-    type OnKilledAccount = ();
-    type SystemWeightInfo = ();
-    type SS58Prefix = SS58Prefix;
-    type OnSetCode = ();
-    type MaxConsumers = frame_support::traits::ConstU32<16>;
-}
-
-pub struct GasConverter;
-impl gear_common::GasToFeeConverter for GasConverter {
-    type Balance = u128;
-}
-
-parameter_types! {
-    pub const BlockGasLimit: u64 = 100_000_000_000;
-}
-
-impl pallet_gear::Config for Test {
-    type Event = Event;
-    type Currency = Balances;
-    type GasConverter = GasConverter;
-    type WeightInfo = ();
-    type BlockGasLimit = BlockGasLimit;
-    type DebugInfo = pallet_gear_debug::Pallet<Test>;
-}
-
-impl pallet_gear_debug::Config for Test {
-    type Event = Event;
-    type WeightInfo = ();
-}
-
-pub struct FixedBlockAuthor;
-
-impl FindAuthor<u64> for FixedBlockAuthor {
-    fn find_author<'a, I>(_digests: I) -> Option<u64>
-    where
-        I: 'a + IntoIterator<Item = (sp_runtime::ConsensusEngineId, &'a [u8])>,
-    {
-        Some(BLOCK_AUTHOR)
-    }
-}
-
-impl pallet_authorship::Config for Test {
-    type FindAuthor = FixedBlockAuthor;
-    type UncleGenerations = ();
-    type FilterUncle = ();
-    type EventHandler = ();
-}
-
-parameter_types! {
-    pub const MinimumPeriod: u64 = 500;
-}
-
-impl pallet_timestamp::Config for Test {
-    type Moment = u64;
-    type OnTimestampSet = ();
-    type MinimumPeriod = MinimumPeriod;
-    type WeightInfo = ();
-}
-
-// Build genesis storage according to the mock runtime.
-pub fn new_test_ext() -> sp_io::TestExternalities {
-    let mut t = system::GenesisConfig::default()
-        .build_storage::<Test>()
-        .unwrap();
-
-    pallet_balances::GenesisConfig::<Test> {
-        balances: vec![
-            (
-                USER_1,
-                (<Test as pallet_gear::Config>::BlockGasLimit::get() * 10) as u128,
-            ),
-            (USER_2, 100_000_000_u128),
-            (USER_3, 300_000_000_u128),
-            (LOW_BALANCE_USER, 2_u128),
-            (BLOCK_AUTHOR, 1_u128),
-        ],
-    }
-    .assimilate_storage(&mut t)
-    .unwrap();
-
-    let mut ext = sp_io::TestExternalities::new(t);
-    ext.execute_with(|| System::set_block_number(1));
-    ext
-}
-
-pub fn run_to_block(n: u64, remaining_weight: Option<u64>) {
-    while System::block_number() < n {
-        System::on_finalize(System::block_number());
-        System::set_block_number(System::block_number() + 1);
-        System::on_initialize(System::block_number());
-        Gear::on_initialize(System::block_number());
-        let remaining_weight =
-            remaining_weight.unwrap_or(<Test as pallet_gear::Config>::BlockGasLimit::get());
-        Gear::on_idle(System::block_number(), remaining_weight);
-    }
-}
